feat: move connectDB function to utility module for better organization
